+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- (VANAGON-85) Added the `clear_provisioning` method to the platform dsl to clear 
+the provisioning command array. 
 
 ## [0.21.0] - released 2021-04-15
 ### Added